+++ ark_nlp/model/rc/pure_bert/predictor.py
