+++ ark_nlp/model/tm/unsupervised_simcse/__init__.py
@@ -1,6 +1,6 @@
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_dataset import UnsupervisedSimCSEDataset
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_dataset import UnsupervisedSimCSEDataset as UnsupSimCSEDataset
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_dataset import UnsupervisedSimCSEDataset as Dataset
+from ark_nlp.model.tm.unsupervised_simcse.dataset import UnsupervisedSimCSEDataset
+from ark_nlp.model.tm.unsupervised_simcse.dataset import UnsupervisedSimCSEDataset as UnsupSimCSEDataset
+from ark_nlp.model.tm.unsupervised_simcse.dataset import UnsupervisedSimCSEDataset as Dataset
 
 from ark_nlp.processor.tokenizer.transformer import SentenceTokenizer as Tokenizer
 from ark_nlp.processor.tokenizer.transformer import SentenceTokenizer as UnsupSimCSETokenizer
@@ -18,10 +18,10 @@ from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_unsup_simcse_optimizer
 from ark_nlp.factory.optimizer import get_default_bert_optimizer as get_default_unsupervised_simcse_optimizer
 
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_task import UnsupervisedSimCSETask
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_task import UnsupervisedSimCSETask as Task
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_task import UnsupervisedSimCSETask as UnsupSimCSETask
+from ark_nlp.model.tm.unsupervised_simcse.task import UnsupervisedSimCSETask
+from ark_nlp.model.tm.unsupervised_simcse.task import UnsupervisedSimCSETask as Task
+from ark_nlp.model.tm.unsupervised_simcse.task import UnsupervisedSimCSETask as UnsupSimCSETask
 
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_predictor import UnsupervisedSimCSEPredictor
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_predictor import UnsupervisedSimCSEPredictor as Predictor
-from ark_nlp.model.tm.unsupervised_simcse.unsupervised_simcse_predictor import UnsupervisedSimCSEPredictor as UnsupSimCSEPredictor
+from ark_nlp.model.tm.unsupervised_simcse.predictor import UnsupervisedSimCSEPredictor
+from ark_nlp.model.tm.unsupervised_simcse.predictor import UnsupervisedSimCSEPredictor as Predictor
+from ark_nlp.model.tm.unsupervised_simcse.predictor import UnsupervisedSimCSEPredictor as UnsupSimCSEPredictor
